feat: Move application details from sidesheet to dedicated page (#644)
* refactor: simplify application handling in ApplicationsPage component and enhance ApplicationList interaction

* feat: move application details from sidesheet to dedicated page with two-column layout

- Create new application detail page at [applicationId]/page.tsx
- Implement two-column layout: left for application content/AI evaluation, right for comments
- Update ApplicationList to open applications in new tab instead of sidesheet
- Create ApplicationContent component for displaying application details and status
- Create CommentsSection component for the comments sidebar
- Remove sidesheet logic from main applications page
- Improve responsive design with mobile-friendly header layout
- Hide status actions for approved/rejected applications
- Enhance UI with better spacing and visual hierarchy



* refactor: remove ApplicationDetailSidesheet component and update related imports

- Deleted the ApplicationDetailSidesheet component to streamline the application detail view.
- Updated imports in ApplicationContent and AIEvaluation components to reflect the removal.
- Enhanced the AIEvaluation component by adding a console log for disqualification reasons and improved rendering logic for evaluation data.
- Adjusted the ApplicationContent component to disable the evaluation button based on application status.

This refactor aims to simplify the application detail structure and improve the overall user experience.

* refactor: remove console log from DisqualificationReason component

- Removed the console log statement from the DisqualificationReason component to clean up the code and improve performance.
- This change contributes to a more streamlined and maintainable codebase.

---------
