+++ components/FundingPlatform/ApplicationView/StatusActionButtons.tsx
@@ -21,24 +21,25 @@ const STATUS_TRANSITIONS: Record<ApplicationStatus, StatusTransition[]> = {
       targetStatus: "under_review",
       label: "Start Review",
       variant: "primary",
-      className: "bg-purple-600 hover:bg-purple-700"
+      className: "bg-primary-500 text-white hover:bg-primary-600 dark:text-white dark:bg-primary-900 dark:hover:bg-primary-800"
     }
   ],
   under_review: [
     {
       targetStatus: "revision_requested",
       label: "Request Revision",
-      variant: "secondary"
+      variant: "secondary",
+      className: 'dark:text-white dark:bg-zinc-900 dark:hover:bg-zinc-800'
     },
     {
       targetStatus: "approved",
       label: "Approve",
-      className: "bg-green-600 hover:bg-green-700"
+      className: "border border-emerald-600 text-emerald-700 bg-green-100 hover:bg-green-200 dark:text-white dark:bg-emerald-900 dark:hover:bg-emerald-800"
     },
     {
       targetStatus: "rejected",
       label: "Reject",
-      className: "bg-red-600 hover:bg-red-700"
+      className: "border border-red-600 text-red-700 bg-red-100 hover:bg-red-200 dark:text-white dark:bg-red-900 dark:hover:bg-red-800"
     }
   ],
   revision_requested: [
@@ -46,7 +47,7 @@ const STATUS_TRANSITIONS: Record<ApplicationStatus, StatusTransition[]> = {
       targetStatus: "under_review",
       label: "Review",
       variant: "primary",
-      className: "bg-purple-600 hover:bg-purple-700"
+      className: "bg-primary-500 text-white hover:bg-primary-600 dark:text-white dark:bg-primary-900 dark:hover:bg-primary-800"
     }
   ],
   approved: [],
@@ -69,7 +70,7 @@ const StatusActionButton: FC<StatusActionButtonProps> = ({
     <Button
       onClick={() => onStatusChange(transition.targetStatus)}
       variant={transition.variant}
-      className={`flex-1 ${transition.className || ""}`}
+      className={`flex-1 text-center items-center justify-center ${transition.className || ""}`}
       disabled={disabled}
     >
       {transition.label}
@@ -96,25 +97,23 @@ export const StatusActionButtons: FC<StatusActionButtonsProps> = ({
   }
 
   return (
-    <div className="border-t border-gray-200 dark:border-gray-700 bg-zinc-50 dark:bg-zinc-800 px-4 py-4 sm:px-6">
-      <div className="flex flex-col space-y-2">
-        <div className="flex space-x-3">
-          {availableTransitions.map((transition) => (
-            <StatusActionButton
-              key={transition.targetStatus}
-              transition={transition}
-              onStatusChange={onStatusChange}
-              disabled={isUpdating}
-            />
-          ))}
-        </div>
-
-        {currentStatus === "revision_requested" && (
-          <p className="text-xs text-gray-500 dark:text-gray-400">
-            The applicant can update their submission.
-          </p>
-        )}
+    <div className="flex flex-col space-y-2">
+      <div className="flex space-x-3">
+        {availableTransitions.map((transition) => (
+          <StatusActionButton
+            key={transition.targetStatus}
+            transition={transition}
+            onStatusChange={onStatusChange}
+            disabled={isUpdating}
+          />
+        ))}
       </div>
+
+      {currentStatus === "revision_requested" && (
+        <p className="text-xs text-gray-500 dark:text-gray-400">
+          The applicant can update their submission.
+        </p>
+      )}
     </div>
   );
 };
