+++ components/FundingPlatform/index.ts
@@ -3,4 +3,5 @@ export { ResponsiveApplicationForm } from './ResponsiveApplicationForm';
 // Application Management Components
 export { default as ApplicationListWithAPI } from './ApplicationList/ApplicationListWithAPI';
 export { default as ApplicationList } from './ApplicationList/ApplicationList';
-export { default as ApplicationDetailSidesheet } from './ApplicationView/ApplicationDetailSidesheet'; 
+export { default as ApplicationContent } from './ApplicationView/ApplicationContent';
+export { default as CommentsSection } from './ApplicationView/CommentsSection';
